Merge tag instead of release branch to main for internal releases (#50)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1201899738287924/task/1210994795858971

This change updates TagReleaseAction so that for internal releases it merges the freshly created tag to the base branch,
and not the release branch. This prevents from accidentally merging new unreleased commits added to the release branch
after the internal bump workflow had started.
The public release workflow works the same and it deletes the release branch.