Add support for release branch freezing and unfreezing (#56)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1201899738287924/task/1211100722861854?focus=true

This change adds a new freeze_release_branch_action that is called by the “promote”
workflow on Fridays. The branch is then unfrozen by tag_release_action when run for
a public release (if a branch wasn’t frozen, it handles it gracefully - this will be the case
for hotfixes). Internal release bump workflows run validate_internal_release_bump_action
that now starts with a check for a frozen branch, and if that fails, the workflow is stopped.